Due to differently configured file upload paths in different Linux dist I have now explicetly set the php upload_tmp_dir as a php_admin_value in the vhosts container. Also changed the default ZPanel temp directory to be /var/zpanel/temp and updated install scripts/SQL scripts as required for that change.
This commit also fixes the issue with the openbasedir functionality, it was an incorrect character (the path seperator) which has now been resolved. Had to change from ';' to ':' have tested on my Linux netbook under Ubuntu 11.04 and all seems to be working fine. - Need to check to be sure that this will be ok on Windows but I assume so. Fix bug #57